# apps/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/one-tap-actions/{restaurantId}": {
      "get": {
        "operationId": "OneTapActionsController_getActions",
        "summary": "Get all one-tap actions for a restaurant",
        "parameters": [
          {
            "name": "restaurantId",
            "required": true,
            "in": "path",
            "description": "Restaurant UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "in_progress",
                "completed",
                "cancelled",
                "expired"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneTapActionListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "one-tap-actions"
        ]
      },
      "post": {
        "operationId": "OneTapActionsController_createAction",
        "summary": "Create a new one-tap action",
        "parameters": [
          {
            "name": "restaurantId",
            "required": true,
            "in": "path",
            "description": "Restaurant UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOneTapActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneTapActionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "one-tap-actions"
        ]
      }
    },
    "/api/v1/one-tap-actions/{restaurantId}/pending": {
      "get": {
        "operationId": "OneTapActionsController_getPendingActions",
        "summary": "Get pending one-tap actions for a restaurant",
        "parameters": [
          {
            "name": "restaurantId",
            "required": true,
            "in": "path",
            "description": "Restaurant UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OneTapActionResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "one-tap-actions"
        ]
      }
    },
    "/api/v1/one-tap-actions/action/{actionId}": {
      "get": {
        "operationId": "OneTapActionsController_getAction",
        "summary": "Get a single one-tap action",
        "parameters": [
          {
            "name": "actionId",
            "required": true,
            "in": "path",
            "description": "Action UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneTapActionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Action not found"
          }
        },
        "tags": [
          "one-tap-actions"
        ]
      },
      "put": {
        "operationId": "OneTapActionsController_updateAction",
        "summary": "Update a one-tap action",
        "parameters": [
          {
            "name": "actionId",
            "required": true,
            "in": "path",
            "description": "Action UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOneTapActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneTapActionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Action not found"
          }
        },
        "tags": [
          "one-tap-actions"
        ]
      },
      "delete": {
        "operationId": "OneTapActionsController_deleteAction",
        "summary": "Delete a one-tap action",
        "parameters": [
          {
            "name": "actionId",
            "required": true,
            "in": "path",
            "description": "Action UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Action deleted"
          },
          "404": {
            "description": "Action not found"
          }
        },
        "tags": [
          "one-tap-actions"
        ]
      }
    },
    "/api/v1/one-tap-actions/action/{actionId}/execute": {
      "post": {
        "operationId": "OneTapActionsController_executeAction",
        "summary": "Execute a one-tap action",
        "parameters": [
          {
            "name": "actionId",
            "required": true,
            "in": "path",
            "description": "Action UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneTapActionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Action not found"
          }
        },
        "tags": [
          "one-tap-actions"
        ]
      }
    },
    "/api/v1/one-tap-actions/action/{actionId}/cancel": {
      "post": {
        "operationId": "OneTapActionsController_cancelAction",
        "summary": "Cancel a one-tap action",
        "parameters": [
          {
            "name": "actionId",
            "required": true,
            "in": "path",
            "description": "Action UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneTapActionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Action not found"
          }
        },
        "tags": [
          "one-tap-actions"
        ]
      }
    },
...
  },
  "info": {
    "title": "WineOps AI API",
    "description": "\n## Overview\n\nReal-time API for Restaurant Wine Inventory & Procurement Management.\n\n### Features\n\n- **Dashboard Aggregation**: Single endpoint for all dashboard data (API Bus pattern)\n- **One-Tap Actions**: Quick approval workflows for managers\n- **Toast POS Integration**: Proxy endpoints for Toast API\n- **Real-time Updates**: WebSocket support for live data sync\n- **Rate Limiting**: Per-user and per-restaurant quotas\n\n### Authentication\n\nAll endpoints (except public ones) require a Bearer token in the Authorization header.\n\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n### Rate Limits\n\n| Endpoint Type | Limit | Window |\n|--------------|-------|--------|\n| Default | 100 | 60s |\n| Auth | 10 | 60s |\n| Upload | 10 | 300s |\n| AI | 20 | 60s |\n\n### WebSocket Events\n\nConnect to `ws://localhost:4000` and join rooms:\n- `restaurant:{restaurantId}` - Restaurant-wide events\n- `user:{userId}` - User-specific events\n\nEvents:\n- `notification` - New notifications\n- `inventory_update` - Inventory changes\n- `order_update` - Order status changes\n- `one_tap_action` - One-tap action updates\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "WineOps Team",
      "url": "https://wineops.ai",
      "email": "support@wineops.ai"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://wineops.ai/license"
    }
  },
...
  "servers": [
    {
...
    },
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "OneTapActionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "restaurantId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "actionType": {
            "type": "string",
            "enum": [
              "low_stock",
              "price_change",
              "delivery_confirm",
              "inequality",
              "vintage_sub",
              "stock_receipt",
              "custom",
              "gmail_send",
              "gmail_contextual"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "actionUrl": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "color": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "completed",
              "cancelled",
              "expired"
            ]
          },
          "relatedWineId": {
            "type": "string"
          },
          "relatedOrderId": {
            "type": "string"
          },
          "relatedProviderId": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "executedAt": {
            "type": "string"
          },
          "executedBy": {
            "type": "string"
          },
          "executionResult": {
            "type": "object"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "restaurantId",
          "actionType",
          "title",
          "priority",
          "color",
          "icon",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "OneTapActionListResponseDto": {
        "type": "object",
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OneTapActionResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "pending": {
            "type": "number"
          },
          "completed": {
            "type": "number"
          }
        },
        "required": [
          "actions",
          "total",
          "pending",
          "completed"
        ]
      },
      "CreateOneTapActionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Action title"
          },
          "description": {
            "type": "string",
            "description": "Action description"
          },
          "actionUrl": {
            "type": "string",
            "description": "Action URL to navigate to"
          },
          "actionType": {
            "type": "string",
            "enum": [
              "low_stock",
              "price_change",
              "delivery_confirm",
              "inequality",
              "vintage_sub",
              "stock_receipt",
              "custom",
              "gmail_send",
              "gmail_contextual"
            ],
            "default": "custom"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "default": "medium"
          },
          "color": {
            "type": "string",
            "description": "Color theme",
            "default": "wine"
          },
          "icon": {
            "type": "string",
            "description": "Icon name",
            "default": "Zap"
          },
          "relatedWineId": {
            "type": "string",
            "description": "Related wine ID"
          },
          "relatedOrderId": {
            "type": "string",
            "description": "Related order ID"
          },
          "relatedProviderId": {
            "type": "string",
            "description": "Related provider ID"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata"
          },
          "expiresAt": {
            "type": "string",
            "description": "Expiration date"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateOneTapActionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Action title"
          },
          "description": {
            "type": "string",
            "description": "Action description"
          },
          "actionUrl": {
            "type": "string",
            "description": "Action URL"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "color": {
            "type": "string",
            "description": "Color theme"
          },
          "icon": {
            "type": "string",
            "description": "Icon name"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "completed",
              "cancelled",
              "expired"
            ]
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata"
          }
        }
      },
      "ExecuteActionDto": {
        "type": "object",
        "properties": {
          "result": {
            "type": "object",
            "description": "Execution result data"
          }
        }
      },
...
    }
  }
}